feat: moved utils to app folder
